requirements.txt
aiogram
base114514
aiohttp
pyyaml
pydantic
deepl-translate